fix: combine `Algorithm.h` and `AlgorithmFactory.h`
