+++ src/iguana/algorithms/AlgorithmFactory.cc
@@ -1,4 +1,4 @@
-#include "AlgorithmFactory.h"
+#include "Algorithm.h"
 
 namespace iguana {
